Merge de refactorización y nueva feature menú de pausa.

+++ funciones.h
@@ -22,5 +22,5 @@ void PiezaRotarDerecha(PiezaActiva* p);
 int8_t PiezaDetectarColision(PiezaActiva* p, matrix* m);
 void PiezaRotarIzquierda(PiezaActiva* p);
 int RecalcularVelocidad(int velInicial, int piezas);
-int VelocidadSegunDificultad(int dificultad);
+uint16_t VelocidadSegunDificultad(int dificultad);
 #endif // FUNCIONES_H_INCLUDED
